updates
merged:

  core-gl-testkit: ccc73bddac91b2f1916f3140739fd56e40f73163
  gentoo-staging: 9292583023a6610905c79a4707db3ac6c63b3a31
  kit-fixups: aaab75412e1483afbece97c58c0ce2eee7a9ab6e

+++ x11-drivers/nvidia-drivers/nvidia-drivers-430.50.ebuild
@@ -112,7 +112,7 @@ NV_OPENCL_VEND_DIR="OpenCL/nvidia"
 NV_X_MODDIR="xorg/modules"
 
 # Maximum supported kernel version in form major.minor
-: "${NV_MAX_KERNEL_VERSION:=4.20}"
+: "${NV_MAX_KERNEL_VERSION:=5.4}"
 
 # Fixups for issues with particular versions of the package.
 nv_do_fixups() {
@@ -225,8 +225,10 @@ nv_install_modprobe() {
 # <dir> <template> <perms> <MODULE:>
 nv_install_vulkan_icd() {
 	nv_use "${4#MODULE:}" || return 0
-	rm -f "nvidia_icd.json" || die
-	sed -e 's:__NV_VK_ICD__:'"${NV_NATIVE_LIBDIR}"'/libGLX_nvidia.so.0:g' nvidia_icd.json.template > nvidia_icd.json || die
+	if [ -e nvidia_icd.json.template ]; then
+		rm -f "nvidia_icd.json" || die
+		sed -e 's:__NV_VK_ICD__:'"${NV_NATIVE_LIBDIR}"'/libGLX_nvidia.so.0:g' nvidia_icd.json.template > nvidia_icd.json || die
+	fi
 	nv_install "$1" "nvidia_icd.json" "$3" "$4"
 }
 
@@ -515,12 +517,10 @@ src_install() {
 		newconfd "${FILESDIR}/nvidia-persistenced.conf" nvidia-persistenced
 	fi
 
-	# If we're not using glvnd support, link nvidia opengl vendor directory into system opengl vendor directory
+	# If we're not using glvnd support, then set up directory expected by eselect opengl: 
 	if ! use_if_iuse glvnd ; then
 		dosym "${NV_NATIVE_LIBDIR}/opengl/nvidia" "${EPREFIX}/usr/lib/opengl/nvidia"
-		dosym "${NV_NATIVE_LIBDIR}" "${NV_NATIVE_LIBDIR}/opengl/nvidia/lib"
 	fi
-
 
 	readme.gentoo_create_doc
